// src/types.d.ts
export type Maybe<T> = T | null;

export interface RegisterInput {
  username: string;

  email: string;

  password: string;
}

// ====================================================
// Types
// ====================================================

export interface Query {
  me?: Maybe<User>;
}

export interface User {
  id: string;

  username: string;

  email: string;
}

export interface Mutation {
  register: RegisterResponse;
}

export interface RegisterResponse {
  errors?: Maybe<Error[]>;
}

export interface Error {
  path: string;

  message: string;
}

// ====================================================
// Arguments
// ====================================================

export interface RegisterMutationArgs {
  input: RegisterInput;
}

import { GraphQLResolveInfo } from "graphql";

export type Resolver<Result, Parent = {}, Context = {}, Args = {}> = (
  parent: Parent,
  args: Args,
  context: Context,
  info: GraphQLResolveInfo
) => Promise<Result> | Result;
// ...
  export interface Resolvers<Context = {}, TypeParent = {}> {
    me?: MeResolver<Maybe<User>, TypeParent, Context>;
  }

  export type MeResolver<R = Maybe<User>, Parent = {}, Context = {}> = Resolver<
    R,
    Parent,
    Context
  >;
// ...
export namespace MutationResolvers {
  export interface Resolvers<Context = {}, TypeParent = {}> {
    register?: RegisterResolver<RegisterResponse, TypeParent, Context>;
  }

  export type RegisterResolver<
    R = RegisterResponse,
    Parent = {},
    Context = {}
  > = Resolver<R, Parent, Context, RegisterArgs>;
  export interface RegisterArgs {
    input: RegisterInput;
  }
}